Groc documentation

+++ src/story-edit/keyboard-deletion.js
@@ -1,9 +1,24 @@
+/*
+# keyboard-deletion
+
+This exports functions that add event handlers to `story-edit/view` that allow
+the user to delete selected passages with the delete key.
+*/
+
 'use strict';
 var $ = require('jquery');
 var confirm = require('../ui/modal/confirm');
 var locale = require('../locale');
 
 module.exports = {
+	
+	/*
+	Adds deletion event handlers.
+
+	@method attach
+	@static
+	@param {`story-edit/view`} parent view
+	*/
 	attach: function(parent) {
 		$(document).on('keyup.keyboardDeletion', function handleDeleteKey(e) {
 			if (e.keyCode !== 46 || $('input:focus, textarea:focus').length !== 0) {
@@ -17,46 +32,52 @@ module.exports = {
 			});
 
 			switch (selected.length) {
-			// Bug out if none are selected
-				case 0: {
-					return;
-				}
+				// Bug out if none are selected.
 
-				// Immediately delete if it's just one passage
-				case 1: {
-					parent.deleteSelectedPassages();
-					break;
-				}
+				case 0:
+				return;
 
-				// Show a confirmation modal if it's more than just 1
-				default: {
+				// Immediately delete the selection if it's just one passage.
 
-					// Set count appropriately
+				case 1:
+				parent.deleteSelectedPassages();
+				break;
 
-					// L10n: This message is always shown with more than one passage.
-					// %d is the number of passages.
-					var message = locale.sayPlural(
+				// Show a confirmation modal if it's more than just 1.
+
+				default:
+				// Set count appropriately.
+
+				// L10n: This message is always shown with more than one passage.
+				// %d is the number of passages.
+				var message = locale.sayPlural(
 					'Are you sure you want to delete this passage?',
 					'Are you sure you want to delete these %d passages? ' +
 					'This cannot be undone.',
-					selected.length);
+					selected.length
+				);
 
-					confirm({
-						content: message,
-						confirmLabel:
-						'<i class="fa fa-trash-o"></i> ' + locale.say('Delete'),
-						confirmClass: 'danger',
-						callback: function(confirmed) {
-							if (confirmed) {
-								parent.deleteSelectedPassages();
-							}
+				confirm({
+					content: message,
+					confirmLabel:
+					'<i class="fa fa-trash-o"></i> ' + locale.say('Delete'),
+					confirmClass: 'danger',
+					callback: function(confirmed) {
+						if (confirmed) {
+							parent.deleteSelectedPassages();
 						}
-					});
-				}
+					}
+				});
 			}
 		});
 	},
 
+	/*
+	Removes deletion event handlers.
+
+	@method detatch
+	@static
+	*/
 	detach: function() {
 		$(document).off('.keyboardDeletion');
 	}
